--- src/components/Header/PageList.js
@@ -1,23 +1,0 @@
-// import { NavigateList, Link } from "./Header.styled"
-// import { useState } from "react";
-
-// export const PageList = () => {
-
-//     const [isOpen, setOpen] = useState(false);
-//     const toggle = () => {
-//         setOpen(!isOpen);
-//     };
-
-//     return (
-//         <NavigateList><Link to="/news" onClick={toggle}>
-//             News
-//         </Link>
-//             <Link to="/notices" onClick={toggle}>
-//                 Find pet
-//             </Link>
-//             <Link to="/friends" onClick={toggle}>
-//                 Navigate
-//             </Link>
-//         </NavigateList>
-//     )
-// }
